image-to-text/service/pyproject.toml
[tool.poetry]
name = "service"
version = "0.1.0"
description = ""
authors = ["Emyr298 <emi.ru298@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^11.1.0"
requests = "^2.32.3"
transformers = "^4.47.1"
torch = "^2.5.1"
bentoml = "^1.3.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
